+++ fastcrypto/src/lib.rs
@@ -26,8 +26,8 @@ pub mod ed25519_tests;
 pub mod secp256k1_recoverable_tests;
 
 #[cfg(test)]
-#[path = "tests/secp256r1_tests.rs"]
-pub mod secp256r1_tests;
+#[path = "tests/secp256r1_recoverable_tests.rs"]
+pub mod secp256r1_recoverable_tests;
 
 #[cfg(test)]
 #[path = "tests/bls12381_tests.rs"]
